// src/utils/enums.ts
export enum UserStatus {
    Active = 1,
    Inactive = 2,
}
export enum UserVerificationStatus {
    Verfied = 1,
    NotVerified = 2,
    Expired = 3,
}
export enum AuthService {
    facebook=1,
    gmail=2,
    local=3
}
export enum IconTypes { 
    iconBased = 1,
    intialBased = 2,
    nameBased = 3
}
export enum IconPosition {
    up = 1,
    down = 2,
    left = 3,
    right = 4 
}